+++ samples/Catalog/examples/AnnotationProcessing.tsx
@@ -12,7 +12,7 @@ import { hideToolbar } from '../helpers/NavigationHelper';
 import { PSPDFKit } from '../helpers/PSPDFKit';
 
 export class AnnotationProcessing extends BaseExampleAutoHidingHeaderComponent {
-  pdfRef: React.RefObject<PSPDFKitView>;
+  pdfRef: React.RefObject<PSPDFKitView | null>;
 
   constructor(props: any) {
     super(props);