+++ samples/Catalog/examples/ProgrammaticFormFilling.tsx
@@ -13,7 +13,7 @@ import { extractFromAssetsIfMissing } from '../helpers/FileSystemHelpers';
 import { PSPDFKit } from '../helpers/PSPDFKit';
 
 export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponent {
-  pdfRef: React.RefObject<PSPDFKitView>;
+  pdfRef: React.RefObject<PSPDFKitView | null>;
 
   constructor(props: any) {
     super(props);