+++ samples/Catalog/examples/SaveAs.tsx
@@ -11,7 +11,7 @@ import { BaseExampleAutoHidingHeaderComponent } from '../helpers/BaseExampleAuto
 import { PSPDFKit } from '../helpers/PSPDFKit';
 
 export class SaveAs extends BaseExampleAutoHidingHeaderComponent {
-  pdfRef: React.RefObject<PSPDFKitView>;
+  pdfRef: React.RefObject<PSPDFKitView | null>;
 
   constructor(props: any) {
     super(props);